+++ lib/screens/home_page/editor_pane/agentic_ui_features/tool_generation/ai_toolgen_widgets.dart
@@ -1,13 +1,10 @@
 import 'dart:convert';
-
 import 'package:apidash/apitoolgen/request_consolidator.dart';
 import 'package:apidash/apitoolgen/tool_templates.dart';
-import 'package:apidash/consts.dart';
 import 'package:apidash/screens/common_widgets/ai/ai_model_selector_button.dart';
 import 'package:apidash/services/agentic_services/agent_caller.dart';
 import 'package:apidash/services/agentic_services/agents/apitool_bodygen.dart';
 import 'package:apidash/services/agentic_services/agents/apitool_funcgen.dart';
-import 'package:apidash/screens/home_page/editor_pane/agentic_ui_elements/ai_ui_designer/ai_ui_desginer_widgets.dart';
 import 'package:apidash/widgets/button_copy.dart';
 import 'package:apidash/widgets/previewer_code.dart';
 import 'package:apidash/widgets/widget_sending.dart';
@@ -16,9 +13,8 @@ import 'package:apidash_design_system/tokens/tokens.dart';
 import 'package:apidash_design_system/widgets/popup_menu.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:genai/agentic_engine/blueprint.dart';
-
 import '../../../../../providers/providers.dart';
+import '../ai_ui_designer/generate_ui_dialog.dart';
 
 class GenerateToolDialog extends ConsumerStatefulWidget {
   final APIDashRequestDescription requestDesc;
