+++ packages/genai/lib/genai.dart
@@ -2,3 +2,4 @@ export 'models/models.dart';
 export 'interface/interface.dart';
 export 'utils/utils.dart';
 export 'widgets/widgets.dart';
+export 'agentic_engine/agentic_engine.dart';
